+++ BitFlyerDotNet.LightningApi/Private/GetBalance.cs
@@ -1,6 +1,6 @@
 ﻿//==============================================================================
-// Copyright (c) 2017-2018 Fiats Inc. All rights reserved.
-// http://www.fiats.asia/
+// Copyright (c) 2017-2019 Fiats Inc. All rights reserved.
+// https://www.fiats.asia/
 //
 
 using Newtonsoft.Json;
@@ -15,10 +15,10 @@ namespace BitFlyerDotNet.LightningApi
         public BfCurrencyCode CurrencyCode { get; private set; }
 
         [JsonProperty(PropertyName = "amount")]
-        public double Amount { get; private set; }
+        public decimal Amount { get; private set; }
 
         [JsonProperty(PropertyName = "available")]
-        public double Available { get; private set; }
+        public decimal Available { get; private set; }
     }
 
     public partial class BitFlyerClient
