+++ BitFlyerDotNet.LightningApi/Private/GetCollateralHistory.cs
@@ -1,6 +1,6 @@
 ﻿//==============================================================================
-// Copyright (c) 2017-2018 Fiats Inc. All rights reserved.
-// http://www.fiats.asia/
+// Copyright (c) 2017-2019 Fiats Inc. All rights reserved.
+// https://www.fiats.asia/
 //
 
 using System;
@@ -19,10 +19,10 @@ namespace BitFlyerDotNet.LightningApi
         public BfCurrencyCode CurrencyCode { get; private set; }
 
         [JsonProperty(PropertyName = "change")]
-        public double Change { get; private set; }
+        public decimal Change { get; private set; }
 
         [JsonProperty(PropertyName = "amount")]
-        public double Amount { get; private set; }
+        public decimal Amount { get; private set; }
 
         [JsonProperty(PropertyName = "reason_code")]
         public string ReasonCode { get; private set; }
