+++ BitFlyerDotNet.LightningApi/Private/GetPositions.cs
@@ -1,6 +1,6 @@
 ﻿//==============================================================================
-// Copyright (c) 2017-2018 Fiats Inc. All rights reserved.
-// http://www.fiats.asia/
+// Copyright (c) 2017-2019 Fiats Inc. All rights reserved.
+// https://www.fiats.asia/
 //
 
 using System;
@@ -20,31 +20,31 @@ namespace BitFlyerDotNet.LightningApi
         public BfTradeSide Side { get; private set; }
 
         [JsonProperty(PropertyName = "price")]
-        public double Price { get; private set; }
+        public decimal Price { get; private set; }
 
         [JsonProperty(PropertyName = "size")]
-        public double Size { get; private set; }
+        public decimal Size { get; private set; }
 
         [JsonProperty(PropertyName = "commission")]
-        public double Commission { get; private set; }
+        public decimal Commission { get; private set; }
 
         [JsonProperty(PropertyName = "swap_point_accumulate")]
-        public double SwapPointAccumulate { get; private set; }
+        public decimal SwapPointAccumulate { get; private set; }
 
         [JsonProperty(PropertyName = "require_collateral")]
-        public double RequireCollateral { get; private set; }
+        public decimal RequireCollateral { get; private set; }
 
         [JsonProperty(PropertyName = "open_date")]
         public DateTime OpenDate { get; private set; }
 
         [JsonProperty(PropertyName = "leverage")]
-        public double Leverage { get; private set; }
+        public decimal Leverage { get; private set; }
 
         [JsonProperty(PropertyName = "pnl")]
-        public double ProfitAndLoss { get; private set; }
+        public decimal ProfitAndLoss { get; private set; }
 
         [JsonProperty(PropertyName = "sfd")]
-        public double SwapForDifference { get; private set; }
+        public decimal SwapForDifference { get; private set; }
 
         public override int GetHashCode()
         {
